CLE-NST-AJ_-_Add_tests_for_Java_date_time - java.time_ tests moved to a separate file + added missing tests
